implement csv parser to convert food data into meal objects

+++ src/main/kotlin/model/Nutrition.kt
@@ -1,11 +1,11 @@
 package org.beijing.model
 
 data class Nutrition(
-    val calories: Double,
-    val totalFat: Double,
-    val sugar: Double,
-    val sodium: Double,
-    val protein: Double,
-    val saturatedFat: Double,
-    val carbohydrates: Double
+    val calories: Float,
+    val totalFat: Float,
+    val sugar: Float,
+    val sodium: Float,
+    val protein: Float,
+    val saturatedFat: Float,
+    val carbohydrates: Float
 )
